refactor: move help functions to utils
