+++ tests-fuzz/src/utils/cluster_info.rs
@@ -12,10 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::time::Duration;
+
+use common_telemetry::info;
+use humantime::parse_duration;
 use snafu::ResultExt;
 use sqlx::database::HasArguments;
-use sqlx::{ColumnIndex, Database, Decode, Encode, Executor, IntoArguments, Type};
+use sqlx::{ColumnIndex, Database, Decode, Encode, Executor, IntoArguments, MySql, Pool, Type};
 
+use super::wait::wait_condition_fn;
 use crate::error::{self, Result};
 
 pub const PEER_TYPE_DATANODE: &str = "DATANODE";
@@ -44,4 +49,41 @@ where
         .fetch_all(e)
         .await
         .context(error::ExecuteQuerySnafu { sql })
+}
+
+/// Waits until all datanodes are online within a specified timeout period.
+///
+/// This function repeatedly checks the status of all datanodes and waits until all of them are online
+/// or the timeout period elapses. A datanode is considered online if its `active_time` is less than 3 seconds.
+pub async fn wait_for_all_datanode_online(greptime: Pool<MySql>, timeout: Duration) {
+    wait_condition_fn(
+        timeout,
+        || {
+            let greptime = greptime.clone();
+            Box::pin(async move {
+                let nodes = fetch_nodes(&greptime)
+                    .await
+                    .unwrap()
+                    .into_iter()
+                    .flat_map(|node| {
+                        if node.peer_type == PEER_TYPE_DATANODE {
+                            Some(node)
+                        } else {
+                            None
+                        }
+                    })
+                    .collect::<Vec<_>>();
+                info!("Waits for all datanode online: {nodes:?}");
+                nodes
+            })
+        },
+        |nodes| {
+            nodes
+                .into_iter()
+                .map(|node| parse_duration(&node.active_time.unwrap()).unwrap())
+                .all(|duration| duration < Duration::from_secs(3))
+        },
+        Duration::from_secs(5),
+    )
+    .await
 }
